frontend refactor. broken styling
