+++ src/frontend/package.json
@@ -4,31 +4,31 @@
   "scripts": {
     "prebuild": "node clean.js",
     "build": "vite build && node embed-script.js",
-    "dev": "vite"
+    "dev": "vite --force"
   },
-  "dependencies": {
+  "devDependencies": {
     "@react-three/drei": "^9.92.5",
     "@react-three/fiber": "^8.15.12",
     "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.2.18",
     "@types/three": "^0.159.0",
     "@vitejs/plugin-react": "^4.2.1",
     "archiver": "^6.0.1",
     "autoprefixer": "^10.4.17",
-    "esbuild-helpers": "^1.0.0",
-    "postcss": "^8.4.21",
+    "postcss": "^8.4.33",
     "prettier": "^3.2.4",
     "r3f-perf": "^7.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rollup-plugin-visualizer": "^5.12.0",
-    "tailwindcss": "^3.2.6",
+    "tailwindcss": "^3.4.1",
     "three": "^0.159.0",
     "typescript": "^4.9.5",
-    "vite": "^5.0.10",
-    "zustand": "^4.4.7"
-  },
-  "devDependencies": {
-    "@types/react-dom": "^18.2.18",
+    "vite": "^5.0.12",
+    "zustand": "^4.4.7",
+    "@typescript-eslint/eslint-plugin": "6.18.1",
+    "@typescript-eslint/parser": "6.18.1",
+    "esbuild-helpers": "^1.0.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
