+++ src/frontend/src/components/viewer/GridHelper.tsx
