+++ src/frontend/src/components/viewer/OrientationGizmo.tsx
